Port the build to Jenkins pipeline (CONJ-4442 and CONJ-4443) (#19)
* Add a Jenkinsfile

The goal was to port the build job to a pipeline with minimal changes
in existing scripts. To that end `jenkins.sh` script is left in and
the job steps call out to its various functions, while still
preserving the ability to run everything by just running `./jenkins.sh`.

* Run kitchen tests on Jenkins

Also add scripts needed to set up the environment in test/aws.